backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
cryptography==45.0.2
cymem==2.0.11
docx2txt==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
preshed==3.0.9
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pypdfium2==4.30.1
python-docx==1.1.2
python-multipart==0.0.20
python-pptx==1.0.2
requests==2.32.3
rich==14.0.0
setuptools==80.8.0
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
thinc==8.3.6
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
XlsxWriter==3.2.3
